Small cleanup and warning removal.

+++ src/HX711.cpp
@@ -119,8 +119,8 @@ long HX711::read() {
 	// state after the sequence completes, insuring that the entire read-and-gain-set
 	// sequence is not interrupted.  The macro has a few minor advantages over bracketing
 	// the sequence between `noInterrupts()` and `interrupts()` calls.
-	// Running without interrupts kills the millis() and micros() function because it disables 
-	// Timer0 overflow interrupts, so any other time estimation is messed up. 
+	// Running without interrupts kills the millis() and micros() function because it disables
+	// Timer0 overflow interrupts, so any other time estimation is messed up.
 
 	#if HAS_ATOMIC_BLOCK
 	ATOMIC_BLOCK(ATOMIC_RESTORESTATE) {
@@ -138,7 +138,7 @@ long HX711::read() {
 	// Disable interrupts.
 	noInterrupts();
 	#endif
-	
+
 		// Wait for the chip to become ready.
   if ( !wait_ready_safe() ){
     Serial.println("HX711:wait_ready_safe:Counted out while waiting for is_ready");
@@ -232,15 +232,15 @@ bool HX711::wait_ready_timeout(unsigned long timeout, unsigned long delay_ms) {
 bool HX711::wait_ready_safe(int retries) {
 	// Wait for the chip to become ready and sync with conversion cycle.
 	// Occasionally the HX711 would return an incorrect reading with a value much higher than expected.
-	// The theory is that there is a race condition between the firmware detecting is_ready() 
-	// and the HX711 starting a new conversion cycle after the firmware staring the shift out. 
+	// The theory is that there is a race condition between the firmware detecting is_ready()
+	// and the HX711 starting a new conversion cycle after the firmware staring the shift out.
 	// To prevent this, this function first detects a conversion happening from the HX711, and then
-	// returns soon after the falling edge. 
+	// returns soon after the falling edge.
 	// This function may take up to one conversion cycle to return (100 ms at 10 SPS or ~11 ms at 80 SPS)
-	// In practice, it was clocked in an Arduino MEGA 2560 to talke less than 1200 micros. 
-	// - This could be off by a factor of 10 due to other code that could mess with the timers. 
-  unsigned long n=0;
-  while ( is_ready() && (n++ < retries) ){;} // Wait for conversion start. 
+	// In practice, it was clocked in an Arduino MEGA 2560 to talke less than 1200 micros.
+	// - This could be off by a factor of 10 due to other code that could mess with the timers.
+  int n = 0;
+  while ( is_ready() && (n++ < retries) ){;} // Wait for conversion start.
   if ( n > retries){
     Serial.println("wait_ready_safe:Counted out");
 	  return false;
